upload files

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: pmorello <pmorello@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/25 10:53:19 by pmorello          #+#    #+#             */
-/*   Updated: 2024/02/18 11:08:23 by pmorello         ###   ########.fr       */
+/*   Updated: 2024/02/19 10:13:22 by pmorello         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_atoi(const char *str)
 {
 	int							i;
 	int							y;
-	unsigned long long int		z;
+	int							z;
 
 	i = 0;
 	y = 1;
